+++ actionpack/lib/action_dispatch/testing/integration.rb
@@ -321,7 +321,9 @@ module ActionDispatch
         end
 
         # Performs the actual request.
-        def process(method, path, params: nil, headers: nil, env: nil, xhr: false)
+        def process(method, path, params: nil, headers: nil, env: nil, xhr: false, as: nil)
+          request_encoder = RequestEncoder.encoder(as)
+
           if path =~ %r{://}
             location = URI.parse(path)
             https! URI::HTTPS === location if location.scheme
@@ -330,14 +332,17 @@ module ActionDispatch
               url_host += ":#{location.port}" if default != location.port
               host! url_host
             end
-            path = location.query ? "#{location.path}?#{location.query}" : location.path
+            path = request_encoder.append_format_to location.path
+            path = location.query ? "#{path}?#{location.query}" : path
+          else
+            path = request_encoder.append_format_to path
           end
 
           hostname, port = host.split(':')
 
           request_env = {
             :method => method,
-            :params => params,
+            :params => request_encoder.encode_params(params),
 
             "SERVER_NAME"     => hostname,
             "SERVER_PORT"     => port || (https? ? "443" : "80"),
@@ -347,7 +352,7 @@ module ActionDispatch
             "REQUEST_URI"    => path,
             "HTTP_HOST"      => host,
             "REMOTE_ADDR"    => remote_addr,
-            "CONTENT_TYPE"   => "application/x-www-form-urlencoded",
+            "CONTENT_TYPE"   => request_encoder.content_type,
             "HTTP_ACCEPT"    => accept
           }
 
@@ -376,6 +381,7 @@ module ActionDispatch
           response = _mock_session.last_response
           @response = ActionDispatch::TestResponse.from_response(response)
           @response.request = @request
+          @response.response_parser = RequestEncoder.parser(@response.content_type)
           @html_document = nil
           @url_options = nil
 
@@ -386,6 +392,56 @@ module ActionDispatch
 
         def build_full_uri(path, env)
           "#{env['rack.url_scheme']}://#{env['SERVER_NAME']}:#{env['SERVER_PORT']}#{path}"
+        end
+
+        class RequestEncoder # :nodoc:
+          @encoders = {}
+
+          attr_reader :response_parser
+
+          def initialize(mime_name, param_encoder, response_parser, url_encoded_form = false)
+            @mime = Mime[mime_name]
+
+            unless @mime
+              raise ArgumentError, "Can't register a request encoder for " \
+                "unregistered MIME Type: #{mime_name}. See `Mime::Type.register`."
+            end
+
+            @url_encoded_form = url_encoded_form
+            @path_format      = ".#{@mime.symbol}" unless @url_encoded_form
+            @response_parser  = response_parser || -> body { body }
+            @param_encoder    = param_encoder   || :"to_#{@mime.symbol}".to_proc
+          end
+
+          def append_format_to(path)
+            path << @path_format unless @url_encoded_form
+            path
+          end
+
+          def content_type
+            @mime.to_s
+          end
+
+          def encode_params(params)
+            @param_encoder.call(params)
+          end
+
+          def self.parser(content_type)
+            mime = Mime::Type.lookup(content_type)
+            encoder(mime ? mime.ref : nil).response_parser
+          end
+
+          def self.encoder(name)
+            @encoders[name] || WWWFormEncoder
+          end
+
+          def self.register_encoder(mime_name, param_encoder: nil, response_parser: nil)
+            @encoders[mime_name] = new(mime_name, param_encoder, response_parser)
+          end
+
+          register_encoder :json, response_parser: -> body { JSON.parse(body) }
+
+          WWWFormEncoder = new(:url_encoded_form, -> params { params }, nil, true)
         end
     end
 
@@ -643,6 +699,39 @@ module ActionDispatch
   #       end
   #   end
   #
+  # You can also test your JSON API easily by setting what the request should
+  # be encoded as:
+  #
+  #   require 'test_helper'
+  #
+  #   class ApiTest < ActionDispatch::IntegrationTest
+  #     test 'creates articles' do
+  #       assert_difference -> { Article.count } do
+  #         post articles_path, params: { article: { title: 'Ahoy!' } }, as: :json
+  #       end
+  #
+  #       assert_response :success
+  #       assert_equal({ id: Arcticle.last.id, title: 'Ahoy!' }, response.parsed_body)
+  #     end
+  #   end
+  #
+  # The `as` option sets the format to JSON, sets the content type to
+  # 'application/json' and encodes the parameters as JSON.
+  #
+  # Calling `parsed_body` on the response parses the response body as what
+  # the last request was encoded as. If the request wasn't encoded `as` something,
+  # it's the same as calling `body`.
+  #
+  # For any custom MIME Types you've registered, you can even add your own encoders with:
+  #
+  #   ActionDispatch::IntegrationTest.register_encoder :wibble,
+  #     param_encoder: -> params { params.to_wibble },
+  #     response_parser: -> body { body }
+  #
+  # Where `param_encoder` defines how the params should be encoded and
+  # `response_parser` defines how the response body should be parsed through
+  # `parsed_body`.
+  #
   # Consult the Rails Testing Guide for more.
 
   class IntegrationTest < ActiveSupport::TestCase
@@ -670,6 +759,10 @@ module ActionDispatch
 
     def document_root_element
       html_document.root
+    end
+
+    def self.register_encoder(*args)
+      Integration::Session::RequestEncoder.register_encoder(*args)
     end
   end
 end