+++ actionview/lib/action_view/renderer/partial_renderer/collection_caching.rb
@@ -15,13 +15,16 @@ module ActionView
         return yield unless cache_collection?
 
         keyed_collection = collection_by_cache_keys
-        partial_cache = collection_cache.read_multi(*keyed_collection.keys)
+        cached_partials = collection_cache.read_multi(*keyed_collection.keys)
 
-        @collection = keyed_collection.reject { |key, _| partial_cache.key?(key) }.values
-        rendered_partials = @collection.any? ? yield.dup : []
+        @collection = keyed_collection.reject { |key, _| cached_partials.key?(key) }.values
+        rendered_partials = @collection.empty? ? [] : yield
 
-        fetch_or_cache_partial(partial_cache, order_by: keyed_collection.each_key) do
-          rendered_partials.shift
+        index = 0
+        keyed_collection.map do |cache_key, _|
+          cached_partials.fetch(cache_key) do
+            rendered_partials[index].tap { index += 1 }
+          end
         end
       end
 
@@ -30,12 +33,7 @@ module ActionView
       end
 
       def automatic_cache_eligible?
-        single_template_render? && !callable_cache_key? &&
-          @template.eligible_for_collection_caching?(as: @options[:as])
-      end
-
-      def single_template_render?
-        @template # Template is only set when a collection renders one template.
+        @template && @template.eligible_for_collection_caching?(as: @options[:as])
       end
 
       def callable_cache_key?
@@ -55,15 +53,10 @@ module ActionView
         key.frozen? ? key.dup : key # #read_multi & #write may require mutability, Dalli 2.6.0.
       end
 
-      def fetch_or_cache_partial(cached_partials, order_by:)
-        cache_options = @options[:cache_options] || @locals[:cache_options] || {}
-
-        order_by.map do |key|
-          cached_partials.fetch(key) do
-            yield.tap do |rendered_partial|
-              collection_cache.write(key, rendered_partial, cache_options)
-            end
-          end
+      def collection_cache_rendered_partial(rendered_partial, key_by)
+        if callable_cache_key?
+          key = expanded_cache_key(@options[:cache].call(key_by))
+          collection_cache.write(key, rendered_partial, @options[:cache_options])
         end
       end
   end